Externalize generation of custom filter functionality
